# src/main/resources/api.yaml
openapi: "3.0.3"
info:
  title: Controle de Ponto API
  version: "1.0"
components:
  schemas:
    Mensagem:
      type: object
      properties:
        mensagem:
          type: string
    Momento:
      type: object
      description: O momento da batida
      properties:
        dataHora:
          description: Data e hora da batida
          type: string
          example: "2018-08-22T08:00:00"
    Registro:
      type: object
      properties:
        dia:
          type: string
          format: date
        horarios:
          type: array
          example: ["08:00:00", "12:00:00", "13:00:00", "18:00:00"]
          items:
            type: string

    Relatorio:
      type: object
      description: Relatório mensal
      properties:
        mes:
          type: string
          example: 2018-08
        horasTrabalhadas:
          type: string
          example: PT69H35M5S
        horasExcedentes:
          type: string
          example: PT25M5S
        horasDevidas:
          type: string
          example: PT0S
        registros:
          type: array
          items:
            $ref: '#/components/schemas/Registro'
paths:
  /v1/batidas:
    post:
      tags: 
        - "Batidas"
      summary: "Bater ponto"
      description: "Registrar um horário da jornada diária de trabalho"
      operationId: "insereBatida"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Momento"
      responses:
        "201":
          description: |
            Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registro"
        "400":
          description: |
            Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mensagem"
              examples:
                Data e hora em formato inválido:
                  value:
                    mensagem: Data e hora em formato inválido
                Campo obrigatório não informado:
                  value:
                    mensagem: Campo obrigatório não informado
        "403":
          description: |
            Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mensagem"
              examples:
                apenas 4 horários por dia:
                  value:
                    mensagem: Apenas 4 horários podem ser registrados por dia
                mínimo 1 hora de almoço:
                  value:
                    mensagem: Deve haver no mínimo 1 hora de almoço
                sábado e domingo não são permitidos:
                  value:
                    mensagem: Sábado e domingo não são permitidos como dia de trabalho
        "409":
          description: |
            Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mensagem"
              examples:
                Horário já registrado:
                  value:
                    mensagem: Horário já registrado
  /v1/folhas-de-ponto/{mes}:
    get:
      tags:
        - Folhas de Ponto
      summary: "Relatório mensal"
      description: "Geração de relatório mensal de usuário."
      operationId: "geraRelatorioMensal"
      parameters:
        - name: mes
          in: path
          required: true
          schema:
            type: string
            example: "2018-08"
      responses:
        "200":
          description: Relatório mensal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relatorio"
        "404":
          description: Relatório não encontrado